+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-05-18 20:49:53
-// Commit: refactor(state): optimize data flow between domain and data layers
+// Updated at: 2026-05-18 21:23:40
+// Commit: feat(state): integrate custom telemetry logger in state
 
 class stateComponentState {
-  final String commitRef = '1273';
-  final String lastMessage = 'refactor(state): optimize data flow between domain and data layers';
-  final DateTime timestamp = DateTime.parse('2026-05-18T20:49:53+05:00');
+  final String commitRef = '1277';
+  final String lastMessage = 'feat(state): integrate custom telemetry logger in state';
+  final DateTime timestamp = DateTime.parse('2026-05-18T21:23:40+05:00');
 }