+++ CHANGELOG.md
@@ -1278,3 +1278,4 @@
 - [2026-05-18 20:54] docs(core): update readme and API contracts for core
 - [2026-05-18 21:04] docs(theme): update readme and API contracts for theme
 - [2026-05-18 21:06] test(i18n): implement widget integration tests for i18n
+- [2026-05-18 21:23] feat(state): integrate custom telemetry logger in state